pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
authors = [{ name = "fluentqa", email = "fluentqa@fluent-qa.com" }]

description = "A Simple Configuration Python Project"
version = "0.1.0"
requires-python = ">3.9"
keywords = [
    "python configuration",
    "pydantic",
    "easy settings management",
    "dynaconf",
    "dotenv",
]
name = "qpyconf"
readme = "README.md"
dependencies = [
    "allure-pytest>=2.15.3",
    "dynaconf>=3.2.5",
    "pydantic>=2.7.4",
    "qpyci",
]


[project.urls]
issue = "https://github.com/fluent-qa/qpyconf/issues"
repository = "https://github.com/fluent-qa/qpyconf"

[project.scripts]
cleanup = "qpyci.commands:cleanup"
cov = "qpyci.ci:coverage"
badge = "qpyci.ci:badge"
check_format = "qpyci.ci:check_format"
ci = "qpyci.commands:ci"


[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
qpyci = { git = "https://github.com/fluent-qa/qpyci.git", rev = "main" }

[tool.hatch.build.targets.wheel]
packages = ["src/qpyconf"]

[dependency-groups]
lint = ["ruff>=0.8.3"]
dev = [
    "allure-pytest>=2.15.3",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "coverage-badge>=1.1.2",
]

[tool.ruff]
line-length = 120
target-version = "py39"



